requirements.txt
# ── Core RL ──────────────────────────────────────────────────────────────────
gymnasium>=1.0.0
stable-baselines3[extra]>=2.3.0

# ── Binance (use python-binance for REST+WebSocket, futures-connector for futures) ─
python-binance>=1.0.19
binance-futures-connector>=4.0.0

# ── Feature Engineering ───────────────────────────────────────────────────────
smartmoneyconcepts>=0.0.27          # SMC: OB, FVG, BOS/CHOCH, swing H/L
arch>=7.0.0                         # GARCH volatility
ta>=0.10.2                          # Extra indicators
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# ── Data ─────────────────────────────────────────────────────────────────────
ccxt>=4.3.0                         # Historical data download (offline warm-up)

# ── Config & Env ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1

# ── Monitoring ───────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.20.0

# ── Training Backends ────────────────────────────────────────────────────────
# Modal (serverless GPU) — install only when using Modal backend
# modal>=0.64.0
#
# Lightning.ai SDK — install only when managing studios programmatically
# lightning-sdk>=0.0.7
#
# Uncomment whichever backend you are using, or install manually:
#   pip install modal           (for Modal)
#   pip install lightning-sdk   (for Lightning.ai programmatic control)

# ── Misc ─────────────────────────────────────────────────────────────────────
tqdm>=4.66.0
tensorboard>=2.16.0
